Show events in a <FlatList />
Bare minimum functionality for showing the events in a list

* The events are not sorted correctly
* The item cells should be styled differently and show some more info
  in a more interesting way

But it does handle logging in and out and refetching accordingly:
Before going to the profile and logging in you'll get the public
events. When logging in the private events are fetched and when logging
out again they are replaced with the public events.